api/requirements.txt
asgiref==3.2.3
blis==0.4.1
catalogue==0.0.8
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
cymem==2.0.3
Django==3.0.3
django-cors-headers==3.2.1
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
fuzzywuzzy==0.17.0
idna==2.9
importlib-metadata==1.5.0
isodate==0.6.0
more-itertools==8.2.0
murmurhash==1.0.2
mysql-connector==2.2.9
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.0
oauthlib==3.1.0
plac==1.1.3
preshed==3.0.2
psutil==5.7.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3
rdflib==4.2.2
requests==2.23.0
six==1.14.0
spacy==2.2.3
SPARQLWrapper==1.8.5
sqlparse==0.3.0
srsly==0.2.0
thinc==7.3.1
tqdm==4.43.0
urllib3==1.25.3
wasabi==0.6.0
zipp==0.6.0
